Basic認証のモデル定義をミドルウェアに置かない(リクエスト処理以外の責務を負わない)
